+++ ploy/tests/test_plain.py
@@ -1,6 +1,7 @@
 from mock import MagicMock, call, patch
 from ploy import Controller
 import os
+import paramiko
 import pytest
 
 
@@ -12,13 +13,12 @@ except NameError:  # pragma: nocover
 
 class TestPlain:
     @pytest.fixture(autouse=True)
-    def setup_ctrl(self, os_execvp_mock, paramiko, sshclient, sshconfig, tempdir):
+    def setup_ctrl(self, os_execvp_mock, sshclient, sshconfig, tempdir):
         import ploy.plain
         self.directory = tempdir.directory
         self.ctrl = Controller(self.directory)
         self.ctrl.plugins = {
             'plain': ploy.plain.plugin}
-        self.paramiko = paramiko
         self.ssh_client_mock = sshclient
         self.os_execvp_mock = os_execvp_mock
 
@@ -78,7 +78,7 @@ class TestPlain:
             '[plain-instance:foo]',
             'host = localhost',
             'fingerprint = foo']))
-        self.ssh_client_mock().connect.side_effect = self.paramiko.SSHException(
+        self.ssh_client_mock().connect.side_effect = paramiko.SSHException(
             "Fingerprint doesn't match for localhost (got bar, expected foo)")
         with patch('ploy.log') as LogMock:
             with pytest.raises(SystemExit):
@@ -112,22 +112,16 @@ class TestPlain:
             ['ssh', '-o', 'Forwardagent=yes', '-o', 'StrictHostKeyChecking=yes', '-o', 'UserKnownHostsFile=%s' % known_hosts, '-l', 'root', '-p', '22', 'localhost'])
 
 
-@pytest.fixture
-def paramiko():
-    from ploy.common import import_paramiko
-    return import_paramiko()
-
-
 @pytest.yield_fixture
-def sshconfig(paramiko):
-    with patch("%s.SSHConfig" % paramiko.__name__) as ssh_config_mock:
+def sshconfig():
+    with patch("paramiko.SSHConfig") as ssh_config_mock:
         ssh_config_mock().lookup.return_value = {}
         yield ssh_config_mock
 
 
 @pytest.yield_fixture
-def sshclient(paramiko):
-    with patch("%s.SSHClient" % paramiko.__name__) as ssh_client_mock:
+def sshclient():
+    with patch("paramiko.SSHClient") as ssh_client_mock:
         yield ssh_client_mock
 
 
@@ -168,12 +162,12 @@ def test_conn_no_host(instance):
 
 def test_no_fingerprint(instance):
     instance.config['host'] = 'localhost'
-    with pytest.raises(instance.paramiko.SSHException) as e:
+    with pytest.raises(paramiko.SSHException) as e:
         instance.get_ssh_fingerprints()
     assert e.value.args[0] == "No fingerprint set in config."
 
 
-def test_conn_fingerprint_mismatch(instance, paramiko, sshclient):
+def test_conn_fingerprint_mismatch(instance, sshclient):
     instance.config['host'] = 'localhost'
     instance.config['fingerprint'] = 'foo'
     sshclient().connect.side_effect = paramiko.SSHException(
@@ -191,14 +185,14 @@ def test_conn_fingerprint_mismatch(instance, paramiko, sshclient):
         (("port: 22",), {})]
 
 
-def test_ssh_fingerprints_none_set(instance, paramiko):
+def test_ssh_fingerprints_none_set(instance):
     instance.config['host'] = 'localhost'
     with pytest.raises(paramiko.SSHException) as e:
         instance.get_ssh_fingerprints()
     assert str(e.value) == 'No fingerprint set in config.'
 
 
-def test_ssh_fingerprints_ask(instance, paramiko):
+def test_ssh_fingerprints_ask(instance):
     from ploy.common import SSHKeyFingerprintAsk
     instance.config['host'] = 'localhost'
     instance.config['fingerprint'] = 'ask'
@@ -206,7 +200,7 @@ def test_ssh_fingerprints_ask(instance, paramiko):
     assert isinstance(result, SSHKeyFingerprintAsk)
 
 
-def test_ssh_fingerprints_ignore(instance, paramiko):
+def test_ssh_fingerprints_ignore(instance):
     from ploy.common import SSHKeyFingerprintIgnore
     instance.config['host'] = 'localhost'
     instance.config['fingerprint'] = 'ignore'
@@ -214,7 +208,7 @@ def test_ssh_fingerprints_ignore(instance, paramiko):
     assert isinstance(result, SSHKeyFingerprintIgnore)
 
 
-def test_ssh_fingerprints_auto(instance, paramiko):
+def test_ssh_fingerprints_auto(instance):
     from ploy.common import SSHKeyFingerprintInstance
     instance.config['host'] = 'localhost'
     instance.config['fingerprint'] = 'auto'
@@ -222,7 +216,7 @@ def test_ssh_fingerprints_auto(instance, paramiko):
     assert isinstance(result, SSHKeyFingerprintInstance)
 
 
-def test_ssh_fingerprints_fingerprint(instance, paramiko):
+def test_ssh_fingerprints_fingerprint(instance):
     from ploy.common import SSHKeyFingerprint
     instance.config['host'] = 'localhost'
     instance.config['fingerprint'] = 'a6:7f:6a:a5:8a:7c:26:45:46:ca:d9:d9:8c:f2:64:27'
@@ -230,7 +224,7 @@ def test_ssh_fingerprints_fingerprint(instance, paramiko):
     assert isinstance(result, SSHKeyFingerprint)
 
 
-def test_ssh_fingerprints_fingerprint_auto(instance, paramiko):
+def test_ssh_fingerprints_fingerprint_auto(instance):
     from ploy.common import SSHKeyFingerprint, SSHKeyFingerprintInstance
     instance.config['host'] = 'localhost'
     instance.config['fingerprint'] = 'a6:7f:6a:a5:8a:7c:26:45:46:ca:d9:d9:8c:f2:64:27\nauto'
@@ -302,12 +296,12 @@ def test_conn_cached_closed(instance, sshclient):
     assert second_client.method_calls[2] == call.save_host_keys(instance.master.known_hosts)
 
 
-def test_bad_hostkey(instance, paramiko):
+def test_bad_hostkey(instance):
     with open(instance.master.known_hosts, 'w') as f:
         f.write('foo')
     instance.config['host'] = 'localhost'
     instance.config['fingerprint'] = 'foo'
-    with patch("%s.SSHClient.connect" % paramiko.__name__) as connect_mock:
+    with patch("paramiko.SSHClient.connect") as connect_mock:
         connect_mock.side_effect = [
             paramiko.BadHostKeyException(
                 'localhost', paramiko.PKey('bar'), paramiko.PKey('foo')),
@@ -318,20 +312,20 @@ def test_bad_hostkey(instance, paramiko):
         assert f.read() == ''
 
 
-def test_proxycommand(instance, paramiko, sshclient, tempdir):
+def test_proxycommand(instance, sshclient, tempdir):
     with open(instance.master.known_hosts, 'w') as f:
         f.write('foo')
     instance.config['host'] = 'localhost'
     instance.config['fingerprint'] = 'foo'
     instance.config['proxycommand'] = 'nohup {path}/../bin/ploy-ssh {instances[foo].host} -o UserKnownHostsFile={known_hosts}'
-    with patch("%s.ProxyCommand" % paramiko.__name__) as ProxyCommandMock:
+    with patch("paramiko.ProxyCommand") as ProxyCommandMock:
         info = instance.init_ssh_key()
     proxycommand = 'nohup %s/../bin/ploy-ssh localhost -o UserKnownHostsFile=%s' % (tempdir.directory, instance.master.known_hosts)
     assert info['ProxyCommand'] == proxycommand
     assert ProxyCommandMock.call_args_list == [call(proxycommand)]
 
 
-def test_proxycommand_with_instance(ctrl, paramiko, sshclient):
+def test_proxycommand_with_instance(ctrl, sshclient):
     master = ctrl.instances['master']
     instance = ctrl.instances['foo']
     with open(instance.master.known_hosts, 'w') as f:
@@ -339,14 +333,14 @@ def test_proxycommand_with_instance(ctrl, paramiko, sshclient):
     instance.config['host'] = 'localhost'
     instance.config['fingerprint'] = 'foo'
     instance.config['proxycommand'] = instance.proxycommand_with_instance(master)
-    with patch("%s.ProxyCommand" % paramiko.__name__) as ProxyCommandMock:
+    with patch("paramiko.ProxyCommand") as ProxyCommandMock:
         info = instance.init_ssh_key()
     proxycommand = 'nohup ssh -o StrictHostKeyChecking=yes -o UserKnownHostsFile=%s -l root -p 22 example.com -W localhost:22' % instance.master.known_hosts
     assert info['ProxyCommand'] == proxycommand
     assert ProxyCommandMock.call_args_list == [call(proxycommand)]
 
 
-def test_proxycommand_through_instance(ctrl, ployconf, paramiko, sshclient):
+def test_proxycommand_through_instance(ctrl, ployconf, sshclient):
     ployconf.append('[plain-instance:bar]')
     master = ctrl.instances['master']
     instance = ctrl.instances['foo']
@@ -359,7 +353,7 @@ def test_proxycommand_through_instance(ctrl, ployconf, paramiko, sshclient):
     instance2.config['host'] = 'bar.example.com'
     instance2.config['fingerprint'] = 'foo'
     instance2.config['proxycommand'] = instance2.proxycommand_with_instance(instance)
-    with patch("%s.ProxyCommand" % paramiko.__name__) as ProxyCommandMock:
+    with patch("paramiko.ProxyCommand") as ProxyCommandMock:
         info = instance2.init_ssh_key()
     proxycommand = 'nohup ssh -o StrictHostKeyChecking=yes -o UserKnownHostsFile=%s -l root -p 22 example.com -W localhost:22' % instance.master.known_hosts
     proxycommand2 = "nohup ssh -o 'ProxyCommand=%s' -o StrictHostKeyChecking=yes -o UserKnownHostsFile=%s -l root -p 22 localhost -W bar.example.com:22" % (proxycommand, instance.master.known_hosts)
@@ -367,7 +361,7 @@ def test_proxycommand_through_instance(ctrl, ployconf, paramiko, sshclient):
     assert ProxyCommandMock.call_args_list == [call(proxycommand2)]
 
 
-def test_missing_host_key_mismatch(paramiko, sshclient):
+def test_missing_host_key_mismatch(sshclient):
     from ploy.common import SSHKeyFingerprint
     from ploy.plain import ServerHostKeyPolicy
     shkp = ServerHostKeyPolicy(lambda: [SSHKeyFingerprint('SHA256:LCa0a2j/xo/5m0U8HTBBNBNCLXBkg7+g+YpeiGJm564')])  # that's sha256 of 'foo'